Added collectibles to the game
The collectible scene is just an Area3D with a Mesh and a Collision- once the player enters the area, a signal "collected" is emitted, then the object is queue_free() 'd. The collectible scene is just a square that rotates about the y axis and floats up and down using a sin(x) function. Added 2 throughout the map for examples and also moved player spawnpoint to the top of a shelf next to a collectible